Time: 40 ms (5.42%) | Memory: 39.6 MB (5.20%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.cpp
@@ -4,7 +4,7 @@ public:
         long long h=0;
 
         for(int i=0;i<piles.size();i++){
-            h +=ceil((double)piles[i]/(double)n);
+            h +=ceil((double)piles[i]/n);
         }
         return h;
     }
